\name{MCrt-package}
\alias{MCrt-package}
\alias{MCrt}
\docType{package}
\title{
\packageTitle{MCrt}
}
\description{
\packageDescription{MCrt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MCrt}
\packageIndices{MCrt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MCrt}

Maintainer: \packageMaintainer{MCrt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<MCrt>:<MCrt>-package]{<MCrt>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
